feat: 삭제 메세지 이후 삭제 함수 불러오기로 수정
- 이전에는 버튼을 누르면 handleDelete 함수 -> onDeleteSchedule 함수 -> 삭제 메시지에 따라 삭제 처리 순으로 불러오는 로직.
- handleDelete -> 삭제 메시지에 따라 onDeleteSchedule 함수 불러와 삭제 처리 순으로 한 단계 간소화.

+++ src/components/Timetable.jsx
@@ -16,6 +16,8 @@ const Timetable = ({ date, schedules, onDeleteSchedule, onAddSchedule }) => {
   };
 
   const handleDelete = (id) => {
+    if (!window.confirm("정말로 삭제하시겠습니까?")) return;
+
     onDeleteSchedule(id);
   };
